% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{rnet_aggregate_extensive}
\alias{rnet_aggregate_extensive}
\alias{rnet_aggregate_intensive}
\title{Weighted aggregation of matched networks}
\usage{
rnet_aggregate_extensive(x, y, matches, ..., y_len = sf::st_length(y))

rnet_aggregate_intensive(x, y, matches, ..., x_len = sf::st_length(x))
}
\arguments{
\item{x}{data.frame}

\item{y}{sf object}

\item{matches}{result of rnet_match()}

\item{...}{unquoted variable names in y}

\item{y_len}{default \code{sf::st_length(y)} a numeric vector the same length as y containing the length of each linestring in y}
}
\description{
Weighted aggregation of matched networks
}
