% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_match.R
\name{rnet_match}
\alias{rnet_match}
\title{Match two road networks}
\usage{
rnet_match(x, y, dist_tolerance, angle_tolerance, trees = c("xy", "x"))
}
\arguments{
\item{x}{the target of the join}

\item{y}{the features that will be joined to \code{x}}

\item{dist_tolerance}{the maximum distances that each line segment from \code{y} can be away from \code{x}}

\item{angle_tolerance}{the maximum difference in slope between line segments to be considered a match}

\item{trees}{whether to build a spatial index on \code{x} or \code{x} and \code{y}.}
}
\description{
Match two road networks
}
\details{
\code{x} and \code{y} are cast as geoarrow arrays via \code{geoarrow::as_geoarrow_array()}.

Spherical geometries are not supported at this moment. Using them will result
in inaccurate measurements.

With a smaller number of features, building a spatial index on only the \code{x}
geometries will be faster than building a spatial index on both \code{x} and \code{y}.
However, at a larger number of features in \code{y}, building a spatial index
can result in significant performance enhancements.
}
